+++ tests/test_units.py
@@ -5,7 +5,9 @@ from typing import TypeVar
 
 from aind_data_schema.utils.units import Size, SizeValue, create_unit_with_value
 
-ScalarType = TypeVar("ScalarType", float, int)
+from decimal import Decimal
+
+ScalarType = TypeVar("ScalarType", Decimal, int)
 
 
 class UnitsTests(unittest.TestCase):
@@ -20,7 +22,7 @@ class UnitsTests(unittest.TestCase):
 
         self.assertEqual(default, SizeValue(value=10.1, unit=Size.MM))
 
-        ArbitraryValue = create_unit_with_value("ArbitraryValue", float, Size, Size.IN)
+        ArbitraryValue = create_unit_with_value("ArbitraryValue", Decimal, Size, Size.IN)
 
         self.assertIsNotNone(ArbitraryValue(value=10, unit=Size.PX))
